+++ src/jasmine-spec-reporter.js
@@ -25,7 +25,7 @@ function initColors(options) {
   colors.setTheme({
     success: options.colors && options.colors.success ? options.colors.success : 'green',
     failure: options.colors && options.colors.failure ? options.colors.failure : 'red',
-    pending: options.colors && options.colors.pending ? options.colors.pending : 'cyan'
+    pending: options.colors && options.colors.pending ? options.colors.pending : 'yellow'
   });
   colors.enabled = true;
 }